Refactor PlantDetail page
Refactored `src/pages/PlantDetail.tsx` into smaller, more manageable files.  Removed any unused imports or files.
